RBG Progress and Ideas & Java Make File
The README.txt is just to keep track of everything I’ve worked on so
that we’re on the same page when we aren’t working together. Also, I’ve
put in the example for the Java Makefile. It can compile the
EventServer.java, we can add to it later.